requirements/base.txt
django
django-authtools
django-bower
django-grappelli
django-jinja
django-redis-cache
django-solo
factory-boy
fake-factory
psycopg2
purl
